No working but attempt to simplify chunk update logic
