+++ src/mesh/Chunk_Mesh.cpp
@@ -1,84 +1,47 @@
-#pragma once
+#include "Chunk_Mesh.h"
 
-#include <string>
-#include <fstream>
-#include <sstream>
-#include <iostream>
-#include <vector>
-
-#include <GL/glew.h>
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
-
-#include "../Shaders/Shader.h"
-#include "utils/Texture_Loader.h"
+#include "../utils/Logger.h"
+#define LOG(severity, msg) Logger::log("Chunk_Mesh.h", severity, msg)
 
 using namespace std;
 
-struct Vertex
-{
-    // Position
-    glm::vec3 Position;
-    // Normal
-    glm::vec3 Normal;
-    // TexCoords
-    glm::vec2 TexCoords;
-};
-
-class Mesh
-{
-public:
-    /*  Functions  */
-    // Constructor
-    Mesh(){}
-
-    Mesh(vector<Vertex> vertices, vector<GLuint> indices, vector<Texture> textures)
-    {
-        this->vertices = vertices;
-        this->indices = indices;
+Chunk_Mesh::Chunk_Mesh(vector<Block_Face> vertices, vector<Texture> textures) {
+        texturesMutex.lock();
         this->textures = textures;
+        texturesMutex.unlock();
 
-        // Now that we have all the required data, set the vertex buffers and its attribute pointers.
-        this->setupMesh();
+        updateVertices(vertices);
     }
 
-    Mesh(vector<Vertex> vertices, vector<GLuint> indices)
-    {
+    void Chunk_Mesh::doRender(Shader shader) {
+        if (meshMutex.try_lock()) {
+            if (texturesMutex.try_lock()) {
+                Draw(shader);
+                texturesMutex.unlock();
+            }
+            else {
+                LOG(DEBUG, "Texture mutex locked");
+            }
+            meshMutex.unlock();
+        }
+        else {
+            LOG(DEBUG, "Mesh mutex locked");
+        }
+    }
+
+    void Chunk_Mesh::updateVertices(vector<Block_Face> vertices) {
+        meshMutex.lock();
         this->vertices = vertices;
-        this->indices = indices;
 
-        // Now that we have all the required data, set the vertex buffers and its attribute pointers.
-        this->setupMesh();
+        setupMesh();
+        meshMutex.unlock();
     }
 
-    void doRender(Shader shader, std::vector<Texture> textures) {
-        this->textures = textures;
-        
-        Draw(shader);
-    }
-
-    void doRender(Shader shader) {
-        Draw(shader);
-    }
-
-private:
-    /*  Mesh Data  */
-    vector<Vertex> vertices;
-    vector<GLuint> indices;
-    vector<Texture> textures;
-    Texture_Loader* textureLoader;
-
-    /*  Render data  */
-    GLuint VAO, VBO, EBO;
-
-    /*  Functions    */
-    // Initializes all the buffer objects/arrays
-    void setupMesh()
+    void Chunk_Mesh::setupMesh()
     {
         // Create buffers/arrays
         glGenVertexArrays(1, &this->VAO);
         glGenBuffers(1, &this->VBO);
-        glGenBuffers(1, &this->EBO);
 
         glBindVertexArray(this->VAO);
         // Load data into vertex buffers
@@ -86,33 +49,39 @@ private:
         // A great thing about structs is that their memory layout is sequential for all its items.
         // The effect is that we can simply pass a pointer to the struct and it translates perfectly to a glm::vec3/2 array which
         // again translates to 3/2 floats which translates to a byte array.
-        glBufferData(GL_ARRAY_BUFFER, this->vertices.size() * sizeof(Vertex), &this->vertices[0], GL_STATIC_DRAW);
-
-        glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, this->EBO);
-        glBufferData(GL_ELEMENT_ARRAY_BUFFER, this->indices.size() * sizeof(GLuint), &this->indices[0], GL_STATIC_DRAW);
+        glBufferData(GL_ARRAY_BUFFER, this->vertices.size() * sizeof(Block_Face), &this->vertices[0], GL_STATIC_DRAW);
 
         // Set the vertex attribute pointers
         // Vertex Positions
         glEnableVertexAttribArray(0);
-        glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(Vertex), (GLvoid*)0);
+        glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(Block_Face), (GLvoid*)0);
         // Vertex Normals
         glEnableVertexAttribArray(1);
-        glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, sizeof(Vertex), (GLvoid*)offsetof(Vertex, Normal));
+        glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, sizeof(Block_Face), (GLvoid*)offsetof(Block_Face, Normal));
         // Vertex Texture Coords
         glEnableVertexAttribArray(2);
-        glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, sizeof(Vertex), (GLvoid*)offsetof(Vertex, TexCoords));
+        glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, sizeof(Block_Face), (GLvoid*)offsetof(Block_Face, TexCoords));
 
         glBindVertexArray(0);
     }
 
     // Render the mesh
-    void Draw(Shader shader)
+    void Chunk_Mesh::Draw(Shader shader)
     {
+        //LOG(DEBUG, "Draw");
+        shader.Use();
+
         // Bind appropriate textures
         GLuint diffuseNr = 1;
         GLuint specularNr = 1;
 
-        shader.Use();
+        if (textures.size() == 0) {
+            LOG(WARN, "Textures array is empty in Draw");
+        }
+
+        if (vertices.size() == 0) {
+            LOG(WARN, "Textures array is empty in Draw");
+        }
 
         for (GLuint i = 0; i < this->textures.size(); i++)
         {
@@ -143,7 +112,7 @@ private:
 
         // Draw mesh
         glBindVertexArray(this->VAO);
-        glDrawElements(GL_TRIANGLES, this->indices.size(), GL_UNSIGNED_INT, 0);
+        glDrawArrays(GL_POINTS, 0, vertices.size());
         glBindVertexArray(0);
 
         // Always good practice to set everything back to defaults once configured.
@@ -153,4 +122,3 @@ private:
             glBindTexture(GL_TEXTURE_2D, 0);
         }
     }
-};
