+++ src/mesh/Mesh.h
@@ -5,21 +5,17 @@
 #include <sstream>
 #include <iostream>
 #include <vector>
-#include <mutex>
 
 #include <GL/glew.h>
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 
-#include "../Shaders/Shader.h"
-#include "utils/Texture_Loader.h"
-
-#include "utils/Logger.h"
-#define LOG(severity, msg) Logger::log("Chunk_Mesh.h", severity, msg)
+#include "../../Shaders/Shader.h"
+#include "../utils/Texture_Loader.h"
 
 using namespace std;
 
-struct Block_Face
+struct Vertex
 {
     // Position
     glm::vec3 Position;
@@ -29,51 +25,51 @@ struct Block_Face
     glm::vec2 TexCoords;
 };
 
-class Chunk_Mesh
+class Mesh
 {
 public:
     /*  Functions  */
     // Constructor
-    Chunk_Mesh(vector<Block_Face> vertices, vector<Texture> textures) {
-        updateVertices(vertices);
-        
-        texturesMutex.lock();
+    Mesh(){}
+
+    Mesh(vector<Vertex> vertices, vector<GLuint> indices, vector<Texture> textures)
+    {
+        this->vertices = vertices;
+        this->indices = indices;
         this->textures = textures;
-        texturesMutex.unlock();
+
+        // Now that we have all the required data, set the vertex buffers and its attribute pointers.
+        this->setupMesh();
+    }
+
+    Mesh(vector<Vertex> vertices, vector<GLuint> indices)
+    {
+        this->vertices = vertices;
+        this->indices = indices;
+
+        // Now that we have all the required data, set the vertex buffers and its attribute pointers.
+        this->setupMesh();
+    }
+
+    void doRender(Shader shader, std::vector<Texture> textures) {
+        this->textures = textures;
+        
+        Draw(shader);
     }
 
     void doRender(Shader shader) {
-        if (meshMutex.try_lock()) {
-            if (texturesMutex.try_lock()) {
-                Draw(shader);
-                texturesMutex.unlock();
-            }
-            else {
-                LOG(DEBUG, "Texture mutex locked");
-            }
-            meshMutex.unlock();
-        }
-        else {
-            LOG(DEBUG, "Mesh mutex locked");
-        }
-    }
-
-    void updateVertices(vector<Block_Face> vertices) {
-        meshMutex.lock();
-        this->vertices = vertices;
-
-        setupMesh();
-        meshMutex.unlock();
+        Draw(shader);
     }
 
 private:
     /*  Mesh Data  */
-    vector<Block_Face> vertices;
+    vector<Vertex> vertices;
+    vector<GLuint> indices;
     vector<Texture> textures;
-    std::mutex meshMutex, texturesMutex;
+    Texture_Loader* textureLoader;
 
     /*  Render data  */
-    GLuint VAO, VBO;
+    GLuint VAO, VBO, EBO;
 
     /*  Functions    */
     // Initializes all the buffer objects/arrays
@@ -82,6 +78,7 @@ private:
         // Create buffers/arrays
         glGenVertexArrays(1, &this->VAO);
         glGenBuffers(1, &this->VBO);
+        glGenBuffers(1, &this->EBO);
 
         glBindVertexArray(this->VAO);
         // Load data into vertex buffers
@@ -89,18 +86,21 @@ private:
         // A great thing about structs is that their memory layout is sequential for all its items.
         // The effect is that we can simply pass a pointer to the struct and it translates perfectly to a glm::vec3/2 array which
         // again translates to 3/2 floats which translates to a byte array.
-        glBufferData(GL_ARRAY_BUFFER, this->vertices.size() * sizeof(Block_Face), &this->vertices[0], GL_STATIC_DRAW);
+        glBufferData(GL_ARRAY_BUFFER, this->vertices.size() * sizeof(Vertex), &this->vertices[0], GL_STATIC_DRAW);
+
+        glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, this->EBO);
+        glBufferData(GL_ELEMENT_ARRAY_BUFFER, this->indices.size() * sizeof(GLuint), &this->indices[0], GL_STATIC_DRAW);
 
         // Set the vertex attribute pointers
         // Vertex Positions
         glEnableVertexAttribArray(0);
-        glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(Block_Face), (GLvoid*)0);
+        glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(Vertex), (GLvoid*)0);
         // Vertex Normals
         glEnableVertexAttribArray(1);
-        glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, sizeof(Block_Face), (GLvoid*)offsetof(Block_Face, Normal));
+        glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, sizeof(Vertex), (GLvoid*)offsetof(Vertex, Normal));
         // Vertex Texture Coords
         glEnableVertexAttribArray(2);
-        glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, sizeof(Block_Face), (GLvoid*)offsetof(Block_Face, TexCoords));
+        glVertexAttribPointer(2, 2, GL_FLOAT, GL_FALSE, sizeof(Vertex), (GLvoid*)offsetof(Vertex, TexCoords));
 
         glBindVertexArray(0);
     }
@@ -108,11 +108,11 @@ private:
     // Render the mesh
     void Draw(Shader shader)
     {
-        shader.Use();
-
         // Bind appropriate textures
         GLuint diffuseNr = 1;
         GLuint specularNr = 1;
+
+        shader.Use();
 
         for (GLuint i = 0; i < this->textures.size(); i++)
         {
@@ -143,7 +143,7 @@ private:
 
         // Draw mesh
         glBindVertexArray(this->VAO);
-        glDrawArrays(GL_POINTS, 0, vertices.size());
+        glDrawElements(GL_TRIANGLES, this->indices.size(), GL_UNSIGNED_INT, 0);
         glBindVertexArray(0);
 
         // Always good practice to set everything back to defaults once configured.
