+++ src/world/Chunk.h
@@ -14,7 +14,7 @@
 
 #include "block/Block_Consts.h"
 #include "Chunk_Consts.h"
-#include "../Chunk_Mesh.h"
+#include "../mesh/Chunk_Mesh.h"
 #include "block/Block.h"
 #include "../utils/Texture_Loader.h"
 
@@ -32,8 +32,6 @@ public:
 	Chunk();
 	Chunk(Block_Consts* blockConsts, int xPos, int zPos);
 	~Chunk();
-
-	Chunk_Mesh* getChunkMesh();
 
 	void doUpdate(Chunk* chunkXPOS, Chunk* chunkXNEG, Chunk* chunkZPOS, Chunk* chunkZNEG);
 	void doPartialUpdate(Chunk* chunkXPOS, Chunk* chunkXNEG, Chunk* chunkZPOS, Chunk* chunkZNEG);
