+++ src/world/block/Block_Consts.h
@@ -3,10 +3,13 @@
 #include <iostream>
 #include <ostream>
 #include <unordered_map>
+#include <string>
+#include <fstream>
+#include <sstream>
+#include <vector>
 
 #include "glm/glm.hpp"
 
-#include "../../Mesh.h"
 #include "../../utils/Texture_Loader.h"
 
 enum Block_Type {
